+++ sdk/test-utils/recorder/CHANGELOG.md
@@ -19,7 +19,7 @@
   - Clarify error when attempting to redirect a request but the recorder has not been started
   - Forward mismatch error when recording file cannot be found during `start` call in playback mode
   - Add more descriptive message in the case that the test proxy has not been started before running the tests
-
+- Ignore `Accept-Language` header for browsers in Playback mode.
 ## 3.0.0 (2023-03-07)
 
 ### Features Added